+++ modules/ocl/src/canny.cpp
@@ -237,7 +237,7 @@ void canny::calcSobelRowPass_gpu(const oclMat &src, oclMat &dx_buf, oclMat &dy_b
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&dy_buf.step));
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&dy_buf.offset));
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     size_t localThreads[3]  = {16, 16, 1};
     openCLExecuteKernel(clCxt, &imgproc_canny, kernelName, globalThreads, localThreads, args, -1, -1);
 }
@@ -266,7 +266,7 @@ void canny::calcMagnitude_gpu(const oclMat &dx_buf, const oclMat &dy_buf, oclMat
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&mag.step));
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&mag.offset));
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     size_t localThreads[3]  = {16, 16, 1};
 
     const char * build_options = L2Grad ? "-D L2GRAD":"";
@@ -290,7 +290,7 @@ void canny::calcMagnitude_gpu(const oclMat &dx, const oclMat &dy, oclMat &mag, i
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&mag.step));
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&mag.offset));
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     size_t localThreads[3]  = {16, 16, 1};
 
     const char * build_options = L2Grad ? "-D L2GRAD":"";
@@ -321,7 +321,7 @@ void canny::calcMap_gpu(oclMat &dx, oclMat &dy, oclMat &mag, oclMat &map, int ro
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&map.offset));
 
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     String kernelName = "calcMap";
     size_t localThreads[3]  = {16, 16, 1};
 
@@ -342,7 +342,7 @@ void canny::edgesHysteresisLocal_gpu(oclMat &map, oclMat &st1, void *counter, in
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&map.step));
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&map.offset));
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     size_t localThreads[3]  = {16, 16, 1};
 
     openCLExecuteKernel(clCxt, &imgproc_canny, kernelName, globalThreads, localThreads, args, -1, -1);
@@ -397,7 +397,7 @@ void canny::getEdges_gpu(oclMat &map, oclMat &dst, int rows, int cols)
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&dst.step));
     args.push_back( std::make_pair( sizeof(cl_int), (void *)&dst.offset));
 
-    size_t globalThreads[3] = {cols, rows, 1};
+    size_t globalThreads[3] = {static_cast<size_t>(cols), static_cast<size_t>(rows), 1};
     size_t localThreads[3]  = {16, 16, 1};
 
     openCLExecuteKernel(clCxt, &imgproc_canny, kernelName, globalThreads, localThreads, args, -1, -1);
